// src/types/db.d.ts
export interface Account {
  accountType: string | null;
  created: string;
  createdBy: string;
  isGroup: Generated<number | null>;
  lft: number;
  modified: string;
  modifiedBy: string;
  name: string;
  parentAccount: string | null;
  rgt: number;
  rootType: string;
}

export interface AccountingLedgerEntry {
  account: string;
  created: string;
  createdBy: string;
  credit: string | null;
  date: string | null;
  debit: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  party: string | null;
  referenceName: string | null;
  referenceType: string | null;
  reverted: Generated<number | null>;
  reverts: string | null;
}

export interface Address {
  addressDisplay: string | null;
  addressLine1: string;
  addressLine2: string | null;
  city: string;
  country: string;
  created: string;
  createdBy: string;
  emailAddress: string | null;
  fax: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  phone: string | null;
  pos: string | null;
  postalCode: string | null;
  state: string | null;
}

export interface AppliedCouponCodes {
  coupons: string | null;
  idx: number;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
}

export interface Batch {
  created: string;
  createdBy: string;
  expiryDate: string | null;
  item: string | null;
  manufactureDate: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
}

export interface BatchSeries {
  created: string;
  createdBy: string;
  current: number;
  modified: string;
  modifiedBy: string;
  name: string;
  padZeros: Generated<number>;
  start: Generated<number>;
}

export interface ClosingAmounts {
  closingAmount: string | null;
  differenceAmount: string | null;
  expectedAmount: string | null;
  idx: number;
  name: string;
  openingAmount: string | null;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
  paymentMethod: string;
}

export interface ClosingCash {
  count: Generated<number>;
  denomination: string;
  idx: number;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
}

export interface CollectionRulesItems {
  collectionFactor: string | null;
  idx: number;
  minimumTotalSpent: string | null;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
  tierName: string | null;
}

export interface Color {
  created: string;
  createdBy: string;
  hexvalue: string;
  modified: string;
  modifiedBy: string;
  name: string;
}

export interface CouponCode {
  couponName: string;
  created: string;
  createdBy: string;
  isEnabled: Generated<number>;
  maxAmount: string | null;
  maximumUse: Generated<number>;
  minAmount: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  pricingRule: string;
  used: Generated<number>;
  validFrom: string;
  validTo: string;
}

export interface Currency {
  created: string;
  createdBy: string;
  fraction: string | null;
  fractionUnits: number | null;
  modified: string;
  modifiedBy: string;
  name: string;
  smallestValue: string | null;
  symbol: string | null;
}

export interface CustomField {
  default: string | null;
  fieldname: string;
  fieldtype: Generated<string>;
  idx: number;
  isRequired: Generated<number | null>;
  label: string;
  name: string;
  options: string | null;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
  references: string | null;
  section: Generated<string | null>;
  tab: Generated<string | null>;
  target: string | null;
}

export interface CustomForm {
  created: string;
  createdBy: string;
  modified: string;
  modifiedBy: string;
  name: string;
}

export interface DefaultCashDenominations {
  denomination: string;
  idx: number;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
}

export interface ERPNextSyncQueue {
  created: string;
  createdBy: string;
  documentName: string;
  modified: string;
  modifiedBy: string;
  name: string;
  referenceType: string;
}

export interface FetchFromERPNextQueue {
  created: string;
  createdBy: string;
  documentName: string;
  modified: string;
  modifiedBy: string;
  name: string;
  referenceType: string;
}

export interface IntegrationErrorLog {
  created: string;
  createdBy: string;
  data: string | null;
  error: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  spacer: string | null;
}

export interface Item {
  barcode: string | null;
  batchSeries: string | null;
  created: string;
  createdBy: string;
  datafromErp: Generated<number | null>;
  description: string | null;
  expenseAccount: string;
  for: Generated<string>;
  hasBatch: Generated<number | null>;
  hasSerialNumber: Generated<number | null>;
  hsnCode: string | null;
  image: string | null;
  incomeAccount: string;
  itemCode: string | null;
  itemGroup: string | null;
  itemType: Generated<string | null>;
  modified: string;
  modifiedBy: string;
  name: string;
  rate: string | null;
  serialNumberSeries: string | null;
  tax: string | null;
  trackItem: Generated<number | null>;
  unit: Generated<string | null>;
}

export interface ItemEnquiry {
  contact: string | null;
  created: string;
  createdBy: string;
  customer: string | null;
  description: string | null;
  item: string;
  modified: string;
  modifiedBy: string;
  name: string;
  similarProduct: string | null;
}

export interface ItemGroup {
  created: string;
  createdBy: string;
  hsnCode: string | null;
  image: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  tax: string | null;
}

export interface JournalEntry {
  attachment: string | null;
  cancelled: number;
  created: string;
  createdBy: string;
  date: string;
  entryType: string;
  modified: string;
  modifiedBy: string;
  name: string;
  numberSeries: Generated<string>;
  referenceDate: string | null;
  referenceNumber: string | null;
  submitted: number;
  userRemark: string | null;
}

export interface JournalEntryAccount {
  account: string;
  credit: string | null;
  debit: string | null;
  idx: number;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
}

export interface Lead {
  address: string | null;
  created: string;
  createdBy: string;
  email: string | null;
  mobile: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  status: Generated<string>;
}

export interface Location {
  address: string | null;
  created: string;
  createdBy: string;
  modified: string;
  modifiedBy: string;
  name: string;
}

export interface LoyaltyPointEntry {
  created: string;
  createdBy: string;
  customer: string;
  expiryDate: string | null;
  invoice: string;
  loyaltyPoints: number | null;
  loyaltyProgram: string;
  loyaltyProgramTier: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  postingDate: string | null;
  purchaseAmount: string;
}

export interface LoyaltyProgram {
  conversionFactor: Generated<string>;
  created: string;
  createdBy: string;
  expenseAccount: string;
  expiryDuration: Generated<number | null>;
  fromDate: string;
  isEnabled: Generated<number>;
  maximumUse: Generated<number | null>;
  modified: string;
  modifiedBy: string;
  name: string;
  toDate: string;
  used: Generated<number | null>;
}

export interface NumberSeries {
  created: string;
  createdBy: string;
  current: number;
  modified: string;
  modifiedBy: string;
  name: string;
  padZeros: Generated<number>;
  referenceType: Generated<string>;
  start: Generated<number>;
}

export interface OpeningAmounts {
  amount: string | null;
  idx: number;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
  paymentMethod: string;
}

export interface OpeningCash {
  count: Generated<number>;
  denomination: string;
  idx: number;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
}

export interface Party {
  address: string | null;
  created: string;
  createdBy: string;
  currency: string | null;
  defaultAccount: string | null;
  email: string | null;
  fromLead: string | null;
  gstin: string | null;
  gstType: Generated<string | null>;
  image: string | null;
  loyaltyPoints: Generated<number | null>;
  loyaltyProgram: string | null;
  modified: string;
  modifiedBy: string;
  name: string;
  outstandingAmount: string | null;
  phone: string | null;
  role: Generated<string>;
}

export interface PatchRun {
  created: string;
  createdBy: string;
  failed: Generated<number | null>;
  modified: string;
  modifiedBy: string;
  name: string;
  version: Generated<string | null>;
}

export interface Payment {
  account: string;
  amount: string;
  attachment: string | null;
  cancelled: number;
  clearanceDate: string | null;
  created: string;
  createdBy: string;
  date: string;
  modified: string;
  modifiedBy: string;
  name: string;
  numberSeries: Generated<string>;
  party: string;
  paymentAccount: string;
  paymentMethod: Generated<string>;
  paymentType: string;
  referenceDate: string | null;
  referenceId: string | null;
  referenceType: string | null;
  submitted: number;
  writeoff: string | null;
}

export interface PaymentFor {
  amount: string;
  idx: number;
  name: string;
  parent: string;
  parentFieldname: string;
  parentSchemaName: string;
  referenceName: string;
  referenceType: string;
}

export interface PaymentMethod {
  account: string | null;
  created: string;
  createdBy: string;
  modified: string;
  modifiedBy: string;
  name: string | null;
  type: string;
}

export interface POSClosingShift {
  closingDate: string | null;
  created: string;
  createdBy: string;
  modified: string;
  modifiedBy: string;
  name: string;
  openingShift: string | null;
}

export interface POSOpeningShift {
  created: string;
  createdBy: string;
  modified: string;
  modifiedBy: string;
  name: string;
  openingDate: string | null;
}
// ...
export interface DB {
  Account: Account;
  AccountingLedgerEntry: AccountingLedgerEntry;
  Address: Address;
  AppliedCouponCodes: AppliedCouponCodes;
  Batch: Batch;
  BatchSeries: BatchSeries;
  ClosingAmounts: ClosingAmounts;
  ClosingCash: ClosingCash;
  CollectionRulesItems: CollectionRulesItems;
  Color: Color;
  CouponCode: CouponCode;
  Currency: Currency;
  CustomField: CustomField;
  CustomForm: CustomForm;
  DefaultCashDenominations: DefaultCashDenominations;
  ERPNextSyncQueue: ERPNextSyncQueue;
  FetchFromERPNextQueue: FetchFromERPNextQueue;
  IntegrationErrorLog: IntegrationErrorLog;
  Item: Item;
  ItemEnquiry: ItemEnquiry;
  ItemGroup: ItemGroup;
  JournalEntry: JournalEntry;
  JournalEntryAccount: JournalEntryAccount;
  Lead: Lead;
  Location: Location;
  LoyaltyPointEntry: LoyaltyPointEntry;
  LoyaltyProgram: LoyaltyProgram;
  NumberSeries: NumberSeries;
  OpeningAmounts: OpeningAmounts;
  OpeningCash: OpeningCash;
  Party: Party;
  PatchRun: PatchRun;
  Payment: Payment;
  PaymentFor: PaymentFor;
  PaymentMethod: PaymentMethod;
  POSClosingShift: POSClosingShift;
  POSOpeningShift: POSOpeningShift;
  POSProfile: POSProfile;
  PriceList: PriceList;
  PriceListItem: PriceListItem;
  PricingRule: PricingRule;
  PricingRuleDetail: PricingRuleDetail;
  PricingRuleItem: PricingRuleItem;
  PrintTemplate: PrintTemplate;
  PurchaseInvoice: PurchaseInvoice;
  PurchaseInvoiceItem: PurchaseInvoiceItem;
  PurchaseReceipt: PurchaseReceipt;
  PurchaseReceiptItem: PurchaseReceiptItem;
  SalesInvoice: SalesInvoice;
  SalesInvoiceItem: SalesInvoiceItem;
  SalesQuote: SalesQuote;
  SalesQuoteItem: SalesQuoteItem;
  SerialNumber: SerialNumber;
  SerialNumberSeries: SerialNumberSeries;
  Shipment: Shipment;
  ShipmentItem: ShipmentItem;
  SingleValue: SingleValue;
  StockLedgerEntry: StockLedgerEntry;
  StockMovement: StockMovement;
  StockMovementItem: StockMovementItem;
  Tax: Tax;
  TaxDetail: TaxDetail;
  TaxSummary: TaxSummary;
  UOM: UOM;
  UOMConversionItem: UOMConversionItem;
}